+++ R/aoa.R
@@ -28,6 +28,7 @@
 #' @param useWeight Logical. Only if a model is given. Weight variables according to importance in the model?
 #' @param LPD Logical. Indicates whether the local point density should be calculated or not.
 #' @param maxLPD numeric or integer. Only if \code{LPD = TRUE}. Number of nearest neighbors to be considered for the calculation of the LPD. Either define a number between 0 and 1 to use a percentage of the number of training samples for the LPD calculation or a whole number larger than 1 and smaller than the number of training samples. CAUTION! If not all training samples are considered, a fitted relationship between LPD and error metric will not make sense (@seealso \code{\link{DItoErrormetric}})
+#' @param verbose Logical. Print progress or not?
 #' @details The Dissimilarity Index (DI), the Local Data Point Density (LPD) and the corresponding Area of Applicability (AOA) are calculated.
 #' If variables are factors, dummy variables are created prior to weighting and distance calculation.
 #'
@@ -146,7 +147,8 @@ aoa <- function(newdata,
                 method="L2",
                 useWeight=TRUE,
                 LPD = FALSE,
-                maxLPD = 1) {
+                maxLPD = 1,
+                verbose = TRUE) {
 
   # handling of different raster formats
   as_stars <- FALSE
@@ -200,8 +202,10 @@ aoa <- function(newdata,
 
   # if not provided, compute train DI
   if(!inherits(trainDI, "trainDI")) {
-    message("No trainDI provided.")
-    trainDI <- trainDI(model, train, variables, weight, CVtest, CVtrain, method, useWeight, LPD)
+    if (verbose) {
+      message("No trainDI provided.")
+    }
+    trainDI <- trainDI(model, train, variables, weight, CVtest, CVtrain, method, useWeight, LPD, verbose)
   }
 
   if (calc_LPD == TRUE) {
@@ -302,11 +306,15 @@ aoa <- function(newdata,
   }
 
   if (calc_LPD == TRUE) {
-    message("Computing DI and LPD of new data...")
+    if (verbose) {
+      message("Computing DI and LPD of new data...")
+    }
 
-    pb <- txtProgressBar(min = 0,
-                         max = nrow(newdataCC),
-                         style = 3)
+    if (verbose) {
+      pb <- txtProgressBar(min = 0,
+                           max = nrow(newdataCC),
+                           style = 3)
+    }
 
     DI_out <- rep(NA, nrow(newdata))
     LPD_out <- rep(NA, nrow(newdata))
@@ -317,24 +325,31 @@ aoa <- function(newdata,
 
       DI_out[okrows[i]] <- knnDI[1]
       LPD_out[okrows[i]] <- sum(knnDI < trainDI$threshold)
-      setTxtProgressBar(pb, i)
+      if (verbose) {
+        setTxtProgressBar(pb, i)
+      }
     }
 
-    close(pb)
+    if (verbose) {
+      close(pb)
+    }
 
     # set maxLPD to max of LPD_out if
     realMaxLPD <- max(LPD_out, na.rm = T)
     if (maxLPD > realMaxLPD) {
-      if (inherits(maxLPD, c("numeric", "integer"))) {
+      if (inherits(maxLPD, c("numeric", "integer")) && verbose) {
         message("Your specified maxLPD is bigger than the real maxLPD of you predictor data.")
       }
-      message(paste("maxLPD is set to", realMaxLPD))
+      if (verbose) {
+        message(paste("maxLPD is set to", realMaxLPD))
+      }
       trainDI$maxLPD <- realMaxLPD
     }
   }
 
-
-  message("Computing AOA...")
+  if (verbose) {
+    message("Computing AOA...")
+  }
 
   #### Create Mask for AOA and return statistics
   if (inherits(out, "SpatRaster")) {
@@ -389,7 +404,9 @@ aoa <- function(newdata,
     result$LPD <- LPD
   }
 
-  message("Finished!")
+  if (verbose) {
+    message("Finished!")
+  }
 
   class(result) <- "aoa"
   return(result)