+++ R/nndm.R
@@ -4,7 +4,7 @@
 #' indices to perform a NNDM LOO CV for map validation.
 #' @author Carles Milà
 #' @param tpoints sf or sfc point object. Contains the training points samples.
-#' @param modeldomain sf polygon object defining the prediction area (see Details).
+#' @param modeldomain sf polygon object or SpatRaster defining the prediction area (see Details).
 #' @param ppoints sf or sfc point object. Contains the target prediction points.
 #' Optional. Alternative to modeldomain (see Details).
 #' @param samplesize numeric. How many points in the modeldomain should be sampled as prediction points?
@@ -34,12 +34,13 @@
 #' When \emph{phi} is set to "max", nearest neighbor distance matching is performed for the entire prediction area. Euclidean distances are used for projected
 #' and non-defined CRS, great circle distances are used for geographic CRS (units in meters).
 #'
-#' The \emph{modeldomain} is a sf polygon that defines the prediction area. The function takes a regular point sample (amount defined by \emph{samplesize)} from the spatial extent.
+#' The \emph{modeldomain} is either a sf polygon that defines the prediction area, or alternatively a SpatRaster out of which a polygon,
+#' transformed into the CRS of the training points, is defined as the outline of all non-NA cells.
+#' Then, the function takes a regular point sample (amount defined by \emph{samplesize)} from the spatial extent.
 #' As an alternative use \emph{ppoints} instead of \emph{modeldomain}, if you have already defined the prediction locations (e.g. raster pixel centroids).
 #' When using either \emph{modeldomain} or \emph{ppoints}, we advise to plot the study area polygon and the training/prediction points as a previous step to ensure they are aligned.
 #'
-#' @note NNDM is a variation of LOOCV and therefore may take a long time for large training data sets.
-#' A k-fold variant will be implemented shortly.
+#' @note NNDM is a variation of LOOCV and therefore may take a long time for large training data sets. See \code{\link{knndm}} for a more efficient k-fold variant of the method.
 #' @seealso \code{\link{geodist}}, \code{\link{knndm}}
 #' @references
 #' \itemize{
@@ -97,8 +98,8 @@
 #' plot(nndm_pred)
 #'
 #' ########################################################################
-#' # Example 3: Real- world example; using a modeldomain instead of previously
-#' # sampled prediction locations
+#' # Example 3: Real- world example; using a SpatRast modeldomain instead
+#' # of previously sampled prediction locations
 #' ########################################################################
 #' \dontrun{
 #' library(sf)
@@ -112,15 +113,11 @@
 #' pts <- st_as_sf(pts,coords=c("Easting","Northing"))
 #' st_crs(pts) <- 26911
 #' studyArea <- rast(system.file("extdata","predictors_2012-03-25.tif",package="CAST"))
-#' studyArea[!is.na(studyArea)] <- 1
-#' studyArea <- as.polygons(studyArea, values = FALSE, na.all = TRUE) |>
-#'     st_as_sf() |>
-#'     st_union()
 #' pts <- st_transform(pts, crs = st_crs(studyArea))
-#' plot(studyArea)
-#' plot(st_geometry(pts), add = TRUE, col = "red")
+#' terra::plot(studyArea[["DEM"]])
+#' terra::plot(vect(pts), add = T)
 #'
-#' nndm_folds <- nndm(pts, modeldomain= studyArea)
+#' nndm_folds <- nndm(pts, modeldomain = studyArea)
 #' plot(nndm_folds)
 #'
 #' #use for cross-validation:
@@ -143,12 +140,36 @@ nndm <- function(tpoints, modeldomain = NULL, ppoints = NULL, samplesize = 1000,
 
   # create sample points from modeldomain
   if(is.null(ppoints)&!is.null(modeldomain)){
+
+    # Check modeldomain is indeed a sf/SpatRaster
+    if(!any(c("sfc", "sf", "SpatRaster") %in% class(modeldomain))){
+      stop("modeldomain must be a sf/sfc object or a 'SpatRaster' object.")
+    }
+
+    # If modeldomain is a SpatRaster, transform into polygon
+    if(any(class(modeldomain) == "SpatRaster")){
+      modeldomain[!is.na(modeldomain)] <- 1
+      modeldomain <- terra::as.polygons(modeldomain, values = FALSE, na.all = TRUE) |>
+        sf::st_as_sf() |>
+        sf::st_union()
+      modeldomain <- sf::st_transform(modeldomain, crs = sf::st_crs(tpoints))
+    }
+
+    # Check modeldomain is indeed a polygon sf
+    if(!any(class(sf::st_geometry(modeldomain)) %in% c("sfc_POLYGON", "sfc_MULTIPOLYGON"))){
+      stop("modeldomain must be a sf/sfc polygon object.")
+    }
+
+    # Check whether modeldomain has the same crs as tpoints
     if(!identical(sf::st_crs(tpoints), sf::st_crs(modeldomain))){
       stop("tpoints and modeldomain must have the same CRS")
     }
+
+    # We sample
     message(paste0(samplesize, " prediction points are sampled from the modeldomain"))
     ppoints <- sf::st_sample(x = modeldomain, size = samplesize, type = sampling)
     sf::st_crs(ppoints) <- sf::st_crs(modeldomain)
+
   }else if(!is.null(ppoints)){
     if(!identical(sf::st_crs(tpoints), sf::st_crs(ppoints))){
       stop("tpoints and ppoints must have the same CRS")
@@ -165,6 +186,9 @@ nndm <- function(tpoints, modeldomain = NULL, ppoints = NULL, samplesize = 1000,
     ppoints <- sf::st_sf(geom=ppoints)
   }
 
+  # Input data checks
+  nndm_checks(tpoints, ppoints, phi, min_train)
+
   # if phi==max calculate the range of the size area
   if(phi=="max"){
     xmin <- min(sf::st_coordinates(ppoints)[,1])
@@ -175,9 +199,6 @@ nndm <- function(tpoints, modeldomain = NULL, ppoints = NULL, samplesize = 1000,
     sf::st_crs(p) <- sf::st_crs(ppoints)
     phi <- as.numeric(max(sf::st_distance(p)))
   }
-
-  # Input data checks
-  nndm_checks(tpoints, ppoints, phi, min_train)
 
   # Compute nearest neighbour distances between training and prediction points
   Gij <- sf::st_distance(ppoints, tpoints)
@@ -238,8 +259,8 @@ nndm <- function(tpoints, modeldomain = NULL, ppoints = NULL, samplesize = 1000,
 nndm_checks <- function(tpoints, ppoints, phi, min_train){
 
   # Check for valid range of phi
-  if(phi < 0 | !is.numeric(phi)){
-    stop("phi must be positive.")
+  if(phi < 0 | (!is.numeric(phi) & phi!= "max")){
+    stop("phi must be positive or set to 'max'.")
   }
 
   # min_train must be a single positive numeric