+++ R/trainDI.R
@@ -24,6 +24,7 @@
 #' @param method Character. Method used for distance calculation. Currently euclidean distance (L2) and Mahalanobis distance (MD) are implemented but only L2 is tested. Note that MD takes considerably longer.
 #' @param useWeight Logical. Only if a model is given. Weight variables according to importance in the model?
 #' @param LPD Logical. Indicates whether the local point density should be calculated or not.
+#' @param verbose Logical. Print progress or not?
 #'
 #' @seealso \code{\link{aoa}}
 #' @importFrom graphics boxplot
@@ -111,7 +112,8 @@ trainDI <- function(model = NA,
                     CVtrain = NULL,
                     method="L2",
                     useWeight = TRUE,
-                    LPD = FALSE){
+                    LPD = FALSE,
+                    verbose = TRUE){
 
   # get parameters if they are not provided in function call-----
   if(is.null(train)){train = aoa_get_train(model)}
@@ -192,10 +194,12 @@ trainDI <- function(model = NA,
     S_inv <- MASS::ginv(S)
   }
 
-  message("Computing DI of training data...")
-  pb <- txtProgressBar(min = 0,
-                       max = nrow(train),
-                       style = 3)
+  if (verbose) {
+    message("Computing DI of training data...")
+    pb <- txtProgressBar(min = 0,
+                         max = nrow(train),
+                         style = 3)
+  }
 
   for(i in seq(nrow(train))){
 
@@ -228,9 +232,14 @@ trainDI <- function(model = NA,
     }else{
       trainDist_min <- append(trainDist_min, min(trainDist, na.rm = TRUE))
     }
-    setTxtProgressBar(pb, i)
+    if (verbose) {
+      setTxtProgressBar(pb, i)
+    }
   }
-  close(pb)
+
+  if (verbose) {
+    close(pb)
+  }
   trainDist_avrgmean <- mean(trainDist_avrg,na.rm=TRUE)
 
 
@@ -254,10 +263,12 @@ trainDI <- function(model = NA,
 
   # calculate trainLPD and avrgLPD according to the CV folds
   if (LPD == TRUE) {
-    message("Computing LPD of training data...")
-    pb <- txtProgressBar(min = 0,
-                         max = nrow(train),
-                         style = 3)
+    if (verbose) {
+      message("Computing LPD of training data...")
+      pb <- txtProgressBar(min = 0,
+                           max = nrow(train),
+                           style = 3)
+    }
 
     trainLPD <- c()
     for (j in  seq(nrow(train))) {
@@ -287,10 +298,14 @@ trainDI <- function(model = NA,
       } else {
         trainLPD <- append(trainLPD, sum(DItrainDist[,1] < thres, na.rm = TRUE))
       }
-      setTxtProgressBar(pb, j)
+      if (verbose) {
+        setTxtProgressBar(pb, j)
+      }
     }
 
-    close(pb)
+    if (verbose) {
+      close(pb)
+    }
 
     # Average LPD in trainData
     avrgLPD <- round(mean(trainLPD))